[improvement][headless]Set default query limits of DataSet.

+++ headless/api/src/main/java/com/tencent/supersonic/headless/api/pojo/ItemValueConfig.java
@@ -14,6 +14,6 @@ public class ItemValueConfig {
     private List<String> blackList = new ArrayList<>();
     private List<String> whiteList = new ArrayList<>();
     private List<String> ruleList = new ArrayList<>();
-    private Long limit;
+    private int limit;
     private DateConf dateConf;
 }
